restored redirect on app error (#129)
I just noticed this in the code. It looks like a work-around for testing that I forgot to remove before committing. (The other UI has the same code and redirects as intended.)

In summary, the app _should_ try to redirect to the Business Registry home page on error (after alerting the user).